requirements.txt
# ---------------------------------------------
# requirements.txt
# Indonesian NLP Portfolio — Text Mining J1
# Universitas Udayana — Informatika 2025
# ---------------------------------------------
# Install: pip install -r requirements.txt
# Colab  : !pip install -r requirements.txt

# -- Core Data Science -------------------------
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# -- Machine Learning --------------------------
scikit-learn>=1.3.0

# -- NLP — Indonesian --------------------------
nltk>=3.8.0
PySastrawi>=1.0.0

# -- Deep Learning / Transformers --------------
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.15.0
accelerate>=0.24.0

# -- Visualization -----------------------------
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
plotly>=5.17.0

# -- Utilities ---------------------------------
tqdm>=4.66.0
gdown>=4.7.0

# -- Notebook ----------------------------------
jupyter>=1.0.0
ipywidgets>=8.0.0

# -- Optional: GPU acceleration ----------------
# If using CUDA, install PyTorch with:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
